diet-planner/backend/requirements.txt
fastapi
uvicorn
python-dotenv
pillow
google-generativeai
python-multipart